refactor: collapse 4 distribution files into Distributions.cs (#220)
